+++ atom/browser/net/atom_ssl_config_service.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "atom/browser/atom_ssl_config_service.h"
+#include "atom/browser/net/atom_ssl_config_service.h"
 
 #include <string>
 #include <vector>
