+++ atom/browser/net/atom_ssl_config_service.h
@@ -2,8 +2,8 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#ifndef ATOM_BROWSER_ATOM_SSL_CONFIG_SERVICE_H_
-#define ATOM_BROWSER_ATOM_SSL_CONFIG_SERVICE_H_
+#ifndef ATOM_BROWSER_NET_ATOM_SSL_CONFIG_SERVICE_H_
+#define ATOM_BROWSER_NET_ATOM_SSL_CONFIG_SERVICE_H_
 
 #include "net/ssl/ssl_config_service.h"
 
@@ -25,4 +25,4 @@ class AtomSSLConfigService : public net::SSLConfigService {
 
 }   // namespace atom
 
-#endif  // ATOM_BROWSER_ATOM_SSL_CONFIG_SERVICE_H_
+#endif  // ATOM_BROWSER_NET_ATOM_SSL_CONFIG_SERVICE_H_
